Move theorems about functions
