+++ lisa-sets/src/main/scala/lisa/maths/settheory/functions/FunctionProperties.scala
@@ -1,0 +1,197 @@
+package lisa.maths.settheory.functions
+
+/**
+ * Classes/properties of functions.
+ *
+ * Describes properties of functions such as being injective ([[FunctionPoperties.injective]]) or being invertible
+ * ([[FunctionPoperties.invertibleFunction]]).
+ */
+object FunctionProperties extends lisa.Main {
+  import lisa.maths.settheory.SetTheory.{
+    relation,
+    pair,
+    relationBetween,
+    emptySetIsASubset,
+    cartesianProduct,
+    relationRange,
+    subsetEqualitySymmetry,
+    relationDomain,
+  }
+  import lisa.maths.settheory.functions.Functionals.{
+    functionFromImpliesFunctional,
+    functionFrom,
+    app,
+    functionImpliesRangeSubsetOfCodomain,
+    inRangeImpliesPullbackExists,
+    functionFromImpliesDomainEq,
+  }
+
+  // var defs
+  private val x = variable
+  private val y = variable
+  private val z = variable
+  private val t = variable
+  private val a = variable
+  private val b = variable
+
+  // relation and function symbols
+  private val r = variable
+  private val f = variable
+  private val g = variable
+
+  /**
+   * Surjective (function) --- a function `f: x → y` is surjective iff it
+   * maps to every `b ∈ y` from atleast one `a ∈ x`.
+   *
+   * `surjective(f, x, y) = f ∈ x → y ∧ ∀ b ∈ y. (∃ a ∈ x. f(a) = b)`
+   */
+  val surjective = DEF(f, x, y) --> functionFrom(f, x, y) /\ ∀(b, in(b, y) ==> ∃(a, in(pair(a, b), f)))
+
+  /**
+   * Alias for [[surjective]]
+   */
+  val onto = surjective
+
+  /**
+   * Injective (function) --- a function `f: x → y` is injective iff it maps
+   * to every `b ∈ y` from atmost one `a ∈ x`.
+   *
+   * `injective(f, x, y) = f ∈ x → y ∧ ∀ b ∈ y. (∃ a ∈ x. f(a) = b) ⟹ (∃! a ∈ x. f(a) = b)`
+   */
+  val injective = DEF(f, x, y) --> functionFrom(f, x, y) /\ ∀(b, in(b, y) ==> (∃(a, in(a, x) /\ in(pair(a, b), f)) ==> ∃!(a, in(a, x) /\ in(pair(a, b), f))))
+
+  /**
+   * Alias for [[injective]]
+   */
+  val oneone = injective
+
+  /**
+   * Bijective function --- a function `f: x → y` is bijective iff it is
+   * [[injective]] and [[surjective]].
+   */
+  val bijective = DEF(f, x, y) --> injective(f, x, y) /\ surjective(f, x, y)
+
+  /**
+   * Invertible Function --- a function from `x` to `y` is invertible iff it is
+   * [[bijective]]. See also, [[inverseFunction]]
+   */
+  val invertibleFunction = DEF(f, x, y) --> bijective(f, x, y)
+
+  /**
+   * Inverse Function --- the inverse of a function `f: x → y`, denoted
+   * `f^-1`, is a function from `y` to `x` such that `∀ a ∈ x, b ∈ y.
+   * f(f^-1(b)) = b ∧ f^-1(f(b)) = b`.
+   */
+  val inverseFunctionOf = DEF(g, f, x, y) --> functionFrom(g, y, x) /\ functionFrom(f, x, y) /\ ∀(a, (in(a, y) ==> (a === app(f, app(g, a)))) /\ (in(a, x) ==> (a === app(g, app(f, a)))))
+
+  // val inverseFunctionExistsIfInvertible = makeTHM(
+  //    invertibleFunction(f, x, y) <=> ∃(g, inverseFunctionOf(g, f, x, y))
+  // ) {
+  //   ???
+  // }
+
+  // val inverseFunctionIsUniqueIfItExists = makeTHM(
+  //   ∃(g, inverseFunctionOf(g, f, x, y)) |- ∃!(g, inverseFunctionOf(g, f, x, y))
+  // ) {
+  //   ???
+  // }
+
+  // val inverseFunctionUniqueness = makeTHM(
+  //    ∃!(g, invertibleFunction(f) ==> inverseFunctionOf(g, f, x, y))
+  // ) {
+  //   ???
+  // }
+
+  // val inverseFunction = DEF (f, x, y) --> The(g, invertibleFunction(f) ==> inverseFunctionOf(g, f, x, y))(inverseFunctionUniqueness)
+
+  /**
+   * Theorem --- if a function is [[surjective]], its range is equal to its codomain.
+   */
+  val surjectiveImpliesRangeIsCodomain = Theorem(
+    surjective(f, x, y) |- (y === relationRange(f))
+  ) {
+    have(surjective(f, x, y) |- ∀(b, in(b, y) ==> ∃(a, in(pair(a, b), f)))) by Tautology.from(surjective.definition)
+    val surjDef = thenHave(surjective(f, x, y) |- in(b, y) ==> ∃(a, in(pair(a, b), f))) by InstantiateForall(b)
+    have(∀(t, in(t, relationRange(f)) <=> (∃(a, in(pair(a, t), f))))) by InstantiateForall(relationRange(f))(relationRange.definition of (r -> f))
+    val rangeDef = thenHave(in(b, relationRange(f)) <=> (∃(a, in(pair(a, b), f)))) by InstantiateForall(b)
+
+    have(surjective(f, x, y) |- in(b, y) ==> in(b, relationRange(f))) by Tautology.from(surjDef, rangeDef)
+    thenHave(surjective(f, x, y) |- ∀(b, in(b, y) ==> in(b, relationRange(f)))) by RightForall
+    val surjsub = andThen(Substitution.applySubst(subsetAxiom of (x -> y, y -> relationRange(f))))
+
+    have((surjective(f, x, y), functionFrom(f, x, y)) |- subset(y, relationRange(f)) /\ subset(relationRange(f), y)) by RightAnd(surjsub, functionImpliesRangeSubsetOfCodomain)
+    val funceq = andThen(Substitution.applySubst(subsetEqualitySymmetry of (x -> y, y -> relationRange(f))))
+
+    val surjfunc = have(surjective(f, x, y) |- functionFrom(f, x, y)) by Tautology.from(surjective.definition)
+
+    have(thesis) by Cut(surjfunc, funceq)
+  }
+
+  /**
+   * Theorem --- Cantor's Theorem
+   *
+   * There is no [[surjective]] mapping ([[functionFrom]]) a set to its [[powerSet]].
+   *
+   * In terms of cardinality, it asserts that a set is strictly smaller than
+   * its power set.
+   */
+  val cantorTheorem = Theorem(
+    surjective(f, x, powerSet(x)) |- ()
+  ) {
+    // define y = {z \in x | ! z \in f(z)}
+    val ydef = ∀(t, in(t, y) <=> (in(t, x) /\ !in(t, app(f, t))))
+
+    // y \subseteq x
+    // y \in P(x)
+    have(ydef |- ydef) by Hypothesis
+    thenHave(ydef |- in(t, y) <=> (in(t, x) /\ !in(t, app(f, t)))) by InstantiateForall(t)
+    thenHave(ydef |- in(t, y) ==> in(t, x)) by Weakening
+    thenHave(ydef |- ∀(t, in(t, y) ==> in(t, x))) by RightForall
+    andThen(Substitution.applySubst(subsetAxiom of (x -> y, y -> x)))
+    andThen(Substitution.applySubst(powerAxiom of (x -> y, y -> x)))
+    val yInPower = thenHave(ydef |- in(y, powerSet(x))) by Restate
+
+    // y \in range(f)
+    have(surjective(f, x, powerSet(x)) |- (powerSet(x) === relationRange(f))) by Restate.from(surjectiveImpliesRangeIsCodomain of (y -> powerSet(x)))
+    andThen(Substitution.applySubst(extensionalityAxiom of (x -> powerSet(x), y -> relationRange(f))))
+    val surjRange = thenHave(surjective(f, x, powerSet(x)) |- in(y, powerSet(x)) <=> in(y, relationRange(f))) by InstantiateForall(y)
+    val yInRange = have((ydef, surjective(f, x, powerSet(x))) |- in(y, relationRange(f))) by Tautology.from(yInPower, surjRange)
+
+    // \exists z. z \in x /\ f(z) = y
+    val surjToFunFrom = have(surjective(f, x, powerSet(x)) |- functionFrom(f, x, powerSet(x))) by Tautology.from(surjective.definition of (y -> powerSet(x)))
+    val existsZdom = have((ydef, surjective(f, x, powerSet(x))) |- ∃(z, in(z, relationDomain(f)) /\ (app(f, z) === y))) by Tautology.from(
+      yInRange,
+      surjective.definition of (y -> powerSet(x)),
+      inRangeImpliesPullbackExists of (z -> y),
+      functionFromImpliesFunctional of (y -> powerSet(x))
+    )
+    val xeqdom = thenHave((ydef, surjective(f, x, powerSet(x)), (relationDomain(f) === x)) |- ∃(z, in(z, x) /\ (app(f, z) === y))) by RightSubstEq.withParametersSimple(
+      List((x, relationDomain(f))),
+      lambda(x, ∃(z, in(z, x) /\ (app(f, z) === y)))
+    )
+    val existsZ = have((ydef, surjective(f, x, powerSet(x))) |- ∃(z, in(z, x) /\ (app(f, z) === y))) by Tautology.from(
+      surjective.definition of (y -> powerSet(x)),
+      functionFromImpliesDomainEq of (y -> powerSet(x)),
+      xeqdom
+    )
+
+    // z \in Y <=> z \in x /\ ! z \in f(z)
+    // y = f(z) so z \in f(z) <=> ! z \in f(z)
+    have(ydef |- ydef) by Hypothesis
+    thenHave(ydef |- in(z, y) <=> (in(z, x) /\ !in(z, app(f, z)))) by InstantiateForall(z)
+    thenHave((ydef, in(z, x), (app(f, z) === y)) |- in(z, y) <=> (in(z, x) /\ !in(z, app(f, z)))) by Weakening
+    thenHave((ydef, in(z, x), (app(f, z) === y)) |- in(z, app(f, z)) <=> (in(z, x) /\ !in(z, app(f, z)))) by RightSubstEq.withParametersSimple(
+      List((y, app(f, z))),
+      lambda(y, in(z, y) <=> (in(z, x) /\ !in(z, app(f, z))))
+    )
+    thenHave((ydef, in(z, x) /\ (app(f, z) === y)) |- ()) by Tautology
+    val existsToContra = thenHave((ydef, ∃(z, in(z, x) /\ (app(f, z) === y))) |- ()) by LeftExists
+
+    have((ydef, surjective(f, x, powerSet(x))) |- ()) by Cut(existsZ, existsToContra)
+    val yToContra = thenHave((∃(y, ydef), surjective(f, x, powerSet(x))) |- ()) by LeftExists
+    val yexists = have(∃(y, ydef)) by Restate.from(comprehensionSchema of (z -> x, φ -> lambda(t, !in(t, app(f, t)))))
+
+    have(thesis) by Cut(yexists, yToContra)
+  }
+
+}
