+++ lisa-sets/src/main/scala/lisa/maths/settheory/orderings/PartialOrders.scala
@@ -3,6 +3,7 @@ package lisa.maths.settheory.orderings
 import lisa.automation.settheory.SetTheoryTactics.*
 import lisa.maths.Quantifiers.*
 import lisa.maths.settheory.SetTheory.*
+import lisa.maths.settheory.functions.*
 
 object PartialOrders extends lisa.Main {
