+++ lisa-sets/src/main/scala/lisa/maths/settheory/orderings/Recursion.scala
@@ -3,6 +3,7 @@ package lisa.maths.settheory.orderings
 import lisa.automation.settheory.SetTheoryTactics.*
 import lisa.maths.Quantifiers.*
 import lisa.maths.settheory.SetTheory.*
+import lisa.maths.settheory.functions.*
 import lisa.maths.settheory.orderings.InclusionOrders.*
 import lisa.maths.settheory.orderings.Induction.*
 import lisa.maths.settheory.orderings.PartialOrders.*
